workflow-executor/pyproject.toml
[tool.poetry]
name = "agent-constructor"
version = "0.1.0"
description = "Multi-agent system constructor"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.31.0"
pyyaml = "^6.0.1"
python-dotenv = "^1.0.1"
openai = "^1.12.0"
langchain = "^0.1.0"
langchain-openai = "^0.0.5"
langchain-community = "^0.0.24"
confluent-kafka = "^2.3.0"
sqlalchemy = "^2.0.27"
psycopg2-binary = "^2.9.9"  # Для PostgreSQL

[tool.poetry.group.dev.dependencies]
black = "^24.2.0"
isort = "^5.13.2"
pytest = "^8.0.0"
pytest-cov = "^4.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py39']

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3

[tool.poetry.scripts]
start = "agent_constructor:main" 